Evaluate the estimators in parallel. (#484)
Only support parallel evaluation for the "low_memory" and "cache_preprocessing" fit modes for now.

Use multithreading to evaluate the model in parallel for each estimator. I selected multithreading over multiprocessing because our benchmarking shows that for longer datasets we spend almost all our time in the flash attention kernel, during which time the GIL is released. This allows multithreading to work efficiently, and it is less complex and avoids starting additional processes (which can take a substantial fraction of the inference time).

Ideally `inference.py` would be refactored, but I only did minimal refactoring in this PR.

In addition to the tests added in this PR, these changes are also convered by the new consistency tests for each inference mode: https://github.com/PriorLabs/TabPFN/pull/498 . Unfortunately, GitHub does not support multi-gpu testing on the CI yet, so the testing of the parallelisation is a bit limited.

I simplified the logic for converting the inputs to tensors and setting the dtype in `_prepare_model_inputs()`:
- Always call `torch.as_tensor()`: this only copies if necessary, no need for `if`
- Don't surpress exceptions raised by `X_full.float()`: the comment says this is to avoid overflow errors, but I can't find any case when `.float()` would through an exception, except if `X_full` is complex