Run the tests on merge to main in addition to in PRs. (#519)
To help catch interactions between concurrent PRs, and to make it easier to determine which commit caused a failure.

Also add a "workflow_dispatch" trigger in case we want to run the tests manually (might be useful for flakes). This can only be used by people with write access to the repository, see
https://docs.github.com/en/actions/how-tos/manage-workflow-runs/manually-run-a-workflow