Switch to uv's "--resolution" option to install minimum dependency set. (#535)
Previously we used our own script to generate the minimum dependencies to test the compatibility of our package. This doesn't support environment markers, which is blocking my next PR, and has also previously caused issues:
https://github.com/PriorLabs/TabPFN/commit/534d4c3cf0f8c337492f9fafe123c10fca9d1921#diff-64a02d4ff986c9ed55bfeb17ee7f2882087cd60211ddcce8836459941bb95bf9
uv now has this functionality built in, so replace our script with that.

I select "lowest-direct", because "lowest" also chooses the minimum versions of transitive dependencies, and these often have no minimum version so we get very old versions. This also matches the behaviour of the current script.

Also add minimum versions to the dev dependencies, which makes "lowest-direct" resolution work nicely when testing locally.

This changes the names of the required checks. My plan to force-merge this PR, then change the names of the required checks to the new ones.