Fix potential race condition in InferenceEngineOnDemand. (#543)
If the number of workers used in `fit_preprocessings()` is >1, then the outputs of `fit_preprocessings()` could end up in a different order to the input configs. This would mean the configs could end up being paired with the wrong outputs when returned from `iter_outputs()`.

I don't think this bug could happen in practice because the `n_workers` parameter is not connected, and we always use one worker: https://github.com/PriorLabs/TabPFN/blob/7373d9dd3b13397ebfdba1262baa3ebca5952e61/src/tabpfn/preprocessing.py#L739
It's reasonably likely we might want to use multiple workers in the near future, so I don't remove n_workers.

I updated the tests to catch this race condition, and verified that the test works by disabling the fix, setting the number of workers to >1, and observing that the test fails.